Halve heartbeat interval.
The changes to use a regularly scheduled timer rather than tearing them
down and setting them back up again mean that the following could occur:

- ping/pong occurs
- server sends packet, acked (so `_seenPacket` is true)
- connection dies
- interval timer picks up, notices `_seenPacket` is true, sets it to
  false, continues
- interval timer picks up, finally notices `_seenPacket` is false.

That is, it can take up to two interval cycles to detect that a
connection has gone away.  Accordingly, I halve the `heartbeatInterval`
so that we detect the connection has gone away, in the worst possible
case, in the same amount of time.

`heartbeatTimeout`, being a `setTimeout` based function, does not
require similar adjustment.